requirements.txt
# Core API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.7
pydantic>=2.6.0
python-dotenv>=1.0.0
requests>=2.31.0

# Database
sqlalchemy>=2.0.25
redis>=5.0.1
psycopg2-binary>=2.9.9  # PostgreSQL
pymysql>=1.1.0          # MySQL

# Data Processing
pandas>=2.2.0
openpyxl>=3.1.2
Pillow>=10.2.0

# OCR Engines (Cloud only - local OCR uses subprocess)
google-cloud-vision>=3.7.0
azure-cognitiveservices-vision-computervision>=0.9.0

# AI Providers
openai>=1.12.0
anthropic>=0.18.1
google-generativeai>=0.4.0

# Task Queue
celery>=5.3.6
